Replace utf8 implementation. The utf8 implementation as included had a bug where you could jump past the end of the javascript buffer in certain cases, causing extra bytes to be written to the end of the minified script. I've replaced the utf8 implementation with the version directly from the source https://github.com/JeffBezanson/cutef8 as it doesn't suffer from this problem.

+++ jsmin.c
@@ -25,6 +25,7 @@ SOFTWARE.
 */
 
 #include <stdio.h>
+#include <stdint.h>
 #include "utf8.h"
 #include "jsmin.h"
 
@@ -77,7 +78,7 @@ jsmin_get(jsmin_obj *jmo)
 	int c = jmo->theLookahead;
 	jmo->theLookahead = 0;
 	if (c == 0) {
-		int byte_len = 0;
+		size_t byte_len = 0;
 		c = u8_nextchar(jmo->javascript, &byte_len);
 		jmo->javascript += byte_len;
 		jmo->byte_len = byte_len;
